Removed Initial state

+++ feature_emi/src/main/kotlin/com/allutils/feature_emi/domain/models/EmiDetailsInput.kt
@@ -1,8 +1,8 @@
 package com.allutils.feature_emi.domain.models
 
 data class EmiDetailsInput(
-    val principal: Double? = null,
+    val principal: Int? = null,
     val interest: Double? = null,
-    val tenure: Double? = null,
-    val emi: Double? = null
+    val tenure: Int? = null,
+    val emi: Int? = null
 )
